requirements.txt
appdirs==1.4.4
asgiref==3.2.5
Babel==2.8.0
colorama==0.4.3
confusable-homoglyphs==3.2.0
distlib==0.3.0
Django==3.0.4
django-ajax-image-upload==0.9.4
django-ckeditor==5.9.0
django-ckeditor-staticfiles==4.3.3
django-crispy-forms==1.9.0
django-file-form==2.1.2
django-js-asset==1.2.2
django-phonenumber-field==4.0.0
django-registration==3.1
django-summernote==0.8.11.6
django-taggit==1.2.0
django-tinymce==3.0.2
django-widget-tweaks==1.4.8
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.6.0
packaging==20.4
phonenumbers==8.12.0
Pillow==7.0.0
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pyparsing==2.4.7
pytz==2019.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tox==3.15.1
virtualenv==20.0.21
whitenoise==5.1.0
zipp==3.1.0
